requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.1
python-dotenv==0.21.0

